Create Lab 2
The user can type either a ‘q’ character or a ‘w’ character into the terminal emulator and send it to the
Xplained board. Any other characters will be ignored by the Xplained board. When the Xplained board
receives a ‘q’ character the LED should turn on and the Xplained board should send the message “ON” to
the terminal emulator. When the Xplained board receives a ‘w’ character the LED should turn off and
the Xplained board should send the message “OFF” to the terminal emulator